// src/app/models/character.interface.ts
export interface Character {
  name: string;
  url: string;
  gender: string;
  height: string;
  mass: string;
  hair_color: string;
  eye_color: string;
  birth_year: string;
  homeworld: string;
  film: string[];
  species: string[];
  vehicles: string[];
  starships: string[];
  skin_color: string;
  isFavorite: boolean;
  created: string;
  edited: string;
}